requirements.txt
black
django
djangorestframework
factory_boy
pytest
validators